requirements.txt
Django==4.2.7
Pillow==10.1.0
python-dotenv==1.0.0
requests==2.31.0
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
django-cleanup==8.0.0
django-storages==1.14.2
django-debug-toolbar==4.2.0
whitenoise==6.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-magic==0.4.27 